Add files via upload
Will change code eventually to scrape the website search in conjunction with using the api
For example `/browse-movies/movie_name/all/all/0/featured/0/all`
Why? Because sometimes it's impossible to search a movie using the api. An example would be the movie "X". The api doesn't seem to work well with single letter movie names but if you feed it into the website url and include the year, you will get a result. There's a lot of nuances with this site but there are ways to go about it.